pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gmaps_avoid_swiss"
version = "0.1.5"
description = "A Python package that customizes Google Maps routing to avoid Swiss routes unless the origin, destination, or waypoints include locations within Switzerland."
readme = "README.rst"
authors = [
  {name = "Gregor Bencat", email = "bencat.gregor@gmail.com"}
]
maintainers = [
  {name = "Gregor Bencat", email = "bencat.gregor@gmail.com"}
]
classifiers = [

]
license = {text = "MIT license"}
dependencies = [
  "google-maps-routing",
  "numpy",
  "shapely",
  "polyline"
]

[project.optional-dependencies]
dev = [
    "coverage",  # testing
    "mypy",  # linting
    "pytest",  # testing
    "ruff"  # linting
]

[project.urls]

bugs = "https://github.com/xbencat/gmaps_avoid_swiss/issues"
changelog = "https://github.com/xbencat/gmaps_avoid_swiss/blob/master/changelog.md"
homepage = "https://github.com/xbencat/gmaps_avoid_swiss"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"*" = ["*.*"]




# Mypy
# ----

[tool.mypy]
files = "."

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
allow_untyped_defs = true
disable_error_code = "attr-defined"

[tool.pytest.ini_options]
pythonpath = ["src"]
